backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
cron-descriptor==1.4.0
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
deprecation==2.1.0
distro==1.8.0
Django==4.2.8
django-celery==3.1.17
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-email-verification==0.3.3
django-google-fonts==0.0.3
django-htmx==1.17.2
django-mathfilters==1.0.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
environs==9.5.0
Faker==21.0.0
flower==2.0.1
fonttools==4.47.0
gunicorn==21.2.0
html5lib==1.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
isort==5.12.0
kombu==5.3.4
marshmallow==3.20.1
netaddr==0.9.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sorl-thumbnail==12.10.0
sqlparse==0.4.4
stripe==7.8.1
tinycss2==1.2.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
vine==5.1.0
wcwidth==0.2.12
weasyprint==60.2
webencodings==0.5.1
wrapt==1.16.0
yookassa==3.0.0
zopfli==0.2.3
